Changed the checksum from u16 to u8 like it should be in gps, added a helpful comment to the top of sdcard that basicly says I need to change the whole damn thing

+++ sf_gps_settings.c
@@ -139,7 +139,7 @@ void sf_gps_set_nmea_rate(u8 GLL, u8 RMC, u8 VTG, u8 GGA, u8 GSA, u8 GSV) {
 
 void sf_gps_checksum_calc(char *checksumBuf, char *messageString) {
    char *next = messageString[2];
-   u16 checksum = messageString[1];
+   u8 checksum = messageString[1];
    
    while (*next != '*') {
       checksum ^= (*next);
